Exercices de révision

+++ TP7/src/fr/lelouet/exerciceComplementaire/ListeDeNotes.java
@@ -2,7 +2,7 @@ package fr.lelouet.exerciceComplementaire;
 import java.util.ArrayList;
 
 public class ListeDeNotes extends ArrayList<Float> {
-    public double moyenne(){
-        return stream().mapToDouble(Float::doubleValue).average().orElse(0.0);
+    public float moyenne(){
+        return (float)stream().mapToDouble(Float::doubleValue).average().orElse(0.0);
     }
 }
